Renaming appLastTaskFailure test to appDebugInfo test
